refactor: Removing rust HDF5
Instead of using hdf5-rust, which cause issues on windows, the rust part will create and get the same data, but will return it instead of writing it in the HDF5.

The goal of that is to keep the HDF5 writing part in python, and make the LST processing still fast on rust.