fix: Sort globals file before converting
Array or set in python are not deterministic, so using the order of the files to fill the metadata in HDF5 can result in different results each time.

i.e: The acquisition_time is not the same in all globals files for a specific run, using the first file of the list to fill this data changed each time it was ran, making comparison useless.